monitoring-stack/requirements-supabase.txt
# Supabase dependencies
supabase>=2.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0
python-dateutil>=2.8.2

# Async support (if needed)
aiohttp>=3.8.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-mock>=3.6.1
